+++ CHANGELOG.md
@@ -7,6 +7,11 @@ The format is based on the [KeepAChangeLog] project.
 
 ## Unreleased
 
+### Added
+- [#739] Better error message for providers which return HTTP Error 405 on userinfo
+
+[#739] https://github.com/OpenIDC/pyoidc/pull/739/
+
 ## 1.2.0 [2020-02-05]
 
 ### Fixed